requirements.txt
Flask==2.3.3
Flask-Login==0.6.3
mysql-connector-python==8.1.0
Werkzeug==2.3.6
Jinja2==3.1.2
itsdangerous==2.1.2
click==8.1.7
colorama==0.4.6
